Move getTestEnv to core/environment
