+++ core/index.ts
@@ -21,4 +21,3 @@ export * from './post.js'
 export * from './i18n.js'
 export * from './html.js'
 export * from './slow.js'
-export { getTestEnvironment } from './test/utils.js'
