+++ core/test/utils.ts
@@ -1,9 +1,7 @@
-import { MemoryStore } from '@logux/core'
-import { atom, cleanStores } from 'nanostores'
+import { cleanStores } from 'nanostores'
 import { fail } from 'node:assert'
 
 import {
-  type BaseRoute,
   Category,
   client,
   enableTestTime,
@@ -11,8 +9,10 @@ import {
   fastCategories,
   Feed,
   Filter,
+  getTestEnvironment,
   Post,
   setupEnvironment,
+  setBaseRoute,
   userId
 } from '../index.js'
 
@@ -26,27 +26,6 @@ export function enableClientTest(env: Partial<EnvironmentAndStore> = {}): void {
 export async function cleanClientTest(): Promise<void> {
   cleanStores(fastCategories, Feed, Filter, Category, Post)
   await client.get()?.clean()
-}
-
-const testRouter = atom<BaseRoute | undefined>()
-
-export function setBaseRoute(route: BaseRoute | undefined): void {
-  testRouter.set(route)
-}
-
-export function getTestEnvironment(): EnvironmentAndStore {
-  return {
-    baseRouter: testRouter,
-    errorEvents: { addEventListener() {} },
-    locale: atom('en'),
-    logStoreCreator: () => new MemoryStore(),
-    networkType: () => ({ saveData: undefined, type: undefined }),
-    openRoute: setBaseRoute,
-    persistentEvents: { addEventListener() {}, removeEventListener() {} },
-    persistentStore: {},
-    restartApp: () => {},
-    translationLoader: async () => ({})
-  }
 }
 
 interface PromiseMock<Result> {
